Reworking Session API
Making the session actor to be more reusable. It now takes in a factory
that creates the reader actor. The reader actor needs to take a
ReaderHalf as its arguments. It is also passed a reference to an actor
that can consume BytesAvailable message.

Testing the concept with a simple separator that collect bytes until it
finds a separator byte. This can be made more generic in the future.